scene: implement a simple scenegraph+rendering engine
Implement a single rendering engine that will take care of drawing stuff
to the screen. Use OpenGL for rendering because it's cross platform and
lwjgl is pretty mature.

Implement a scene graph as well, so that our game code can manipulate
the stuff rendered on the screen as 3d objects in space, which maps well
to our human intuition and should thus make our game code easier to
write and understand.

The scene graph takes care of sorting the objects so that they are
rendered back-to-front (on the z axis), so that alpha transparency is
rendered correctly.

As a starter, (to show and test that the rendering engine is working),
implement a `SquareNode` which renders a colored 1x1 square to the
screen.